cjb-backend/pyproject.toml
[project]
name = "cjb-backend"
version = "0.1.0"
description = "FastAPI backend for CJB"
authors = [
    {name = "Nate Farcasin",email = "natefarcasin@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "SQLAlchemy (>=2.0.0,<3.0.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "alembic (>=1.13.0,<2.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
